requirements.txt
asgiref==3.5.1
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
dj-database-url==0.5.0
Django==4.0.4
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-dia==0.4.1
django-dynamic-breadcrumbs==0.4.2
django-environ==0.9.0
django-extensions==3.1.5
django-heroku==0.3.1
django-storages==1.12.3
django-view-breadcrumbs==2.4.1
djangorestframework==3.13.1
graphviz==0.20
gunicorn==20.1.0
idna==3.3
install==1.3.5
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pydot==1.4.2
pydot-ng==2.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
pipenv==2022.10.12
boto3==1.24.93